封车解封车信息落库

+++ bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/newseal/domain/SealVehicles.java
@@ -16,13 +16,13 @@ public class SealVehicles extends DbEntity {
 	private static final long serialVersionUID = 1L;
 
 	 /** 始发站点Code */
-	private Long createSiteCode;
+	private Integer createSiteCode;
 
 	 /** 始发站点名称 */
 	private String createSiteName;
 
 	 /** 目的站点Code */
-	private Long receiveSiteCode;
+	private Integer receiveSiteCode;
 
 	 /** 目的站点名称 */
 	private String receiveSiteName;
@@ -91,7 +91,7 @@ public class SealVehicles extends DbEntity {
 	 * The set method for createSiteCode.
 	 * @param createSiteCode
 	 */
-	public void setCreateSiteCode(Long createSiteCode) {
+	public void setCreateSiteCode(Integer createSiteCode) {
 		this.createSiteCode = createSiteCode;
 	}
 
@@ -99,7 +99,7 @@ public class SealVehicles extends DbEntity {
 	 * The get method for createSiteCode.
 	 * @return this.createSiteCode
 	 */
-	public Long getCreateSiteCode() {
+	public Integer getCreateSiteCode() {
 		return this.createSiteCode;
 	}
 
@@ -123,7 +123,7 @@ public class SealVehicles extends DbEntity {
 	 * The set method for receiveSiteCode.
 	 * @param receiveSiteCode
 	 */
-	public void setReceiveSiteCode(Long receiveSiteCode) {
+	public void setReceiveSiteCode(Integer receiveSiteCode) {
 		this.receiveSiteCode = receiveSiteCode;
 	}
 
@@ -131,7 +131,7 @@ public class SealVehicles extends DbEntity {
 	 * The get method for receiveSiteCode.
 	 * @return this.receiveSiteCode
 	 */
-	public Long getReceiveSiteCode() {
+	public Integer getReceiveSiteCode() {
 		return this.receiveSiteCode;
 	}
